refactor linking
