+++ packages/expo-router/src/views/Directory.tsx
@@ -13,7 +13,7 @@ import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { RouteNode } from "../Route";
 import { useRoutesContext } from "../context";
 import { matchDeepDynamicRouteName, matchFragmentName } from "../matchers";
-import { Link } from "./Link";
+import { Link } from "../link/Link";
 
 const INDENT = 18;
