+++ packages/expo-router/src/views/ErrorBoundary.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Platform, ScrollView, TouchableOpacity } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 
-import { Link } from "./Link";
+import { Link } from "../link/Link";
 import { ErrorBoundaryProps } from "./Try";
 
 export function ErrorBoundary({ error, retry }: ErrorBoundaryProps) {
